Some subject types could belong to multiple orgs.
For some subject types, it might be ok to belong to
multiple orgs. For sure, e.g. adresses, in -debatable- cases, also
`person:Person`
Anyway, this commit, allows this flow, and based on the config, we can
say yes/no for multiple orgs.

However, there is a potential drawback in this implementation: if we need to take into
account both type AND path to determine whether yes/no it may be
shared accross organisations. Let's keep it simple here, and not
support this for now, and first see whether the business requires us
to do so.